custom-element-registry/index.html

<!doctype html>

<meta charset=utf-8>

<title>Custom Elements: CustomElementRegistry interface</title>

<meta name=author title=snuggs href=//github.com/snuggs>
<meta name=assert content='CustomElementRegistry interface must exist'>

<link rel=help href=https://www.w3.org/TR/custom-elements/>
<link rel=help href=https://html.spec.whatwg.org/multipage/custom-elements.html>
<link rel=help href=https://github.com/w3c/web-platform-tests/blob/master/custom-elements/CustomElementRegistry.html>

<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>

<script src=http://www.snuggsi.es></script>
<script>

test (test => {

  assert_true ('define' in customElements, '`define` exists on window.customElements')
  assert_true ('define' in CustomElementRegistry.prototype, '"define" exists on CustomElementRegistry.prototype')

}, 'CustomElementRegistry interface must have define as a method')


test (test => {

  // where is assert_not_throws ?
  customElements.define ('custom-html-element', HTMLElement)

}, 'customElements.define must not throw the constructor is HTMLElement')


test (test => {

    assert_throws ({'name': 'TypeError'}
    , function () { customElements.define ('badname', 1) }
    , 'customElements.define must throw a TypeError when the element interface is a number')

    assert_throws ({'name': 'TypeError'}
    , function () { customElements.define ('badname', '123') }
    , 'customElements.define must throw a TypeError when the element interface is a string')

    assert_throws ({'name': 'TypeError'}
      , function () { customElements.define('badname', {}) }
      , 'customElements.define must throw a TypeError when the element interface is an object')

    assert_throws ({'name': 'TypeError'}
      , function () { customElements.define('badname', []); }
      , 'customElements.define must throw a TypeError when the element interface is an array')

}, 'customElements.define must throw when the element interface is not a constructor')

</script>

<output id=log></output>
